AddfloatPlusSpace

+++ ft_ftoa.c
@@ -6,7 +6,7 @@
 /*   By: lenzo-pe <lenzo-pe@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/28 19:16:33 by lenzo-pe          #+#    #+#             */
-/*   Updated: 2021/03/29 18:01:12 by lenzo-pe         ###   ########.fr       */
+/*   Updated: 2021/03/30 13:05:50 by lenzo-pe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,8 +39,8 @@ char		*ft_ftoa(long double n, size_t len)
 
 	if (!len)
 		n = ft_round(n);
-	left = (unsigned)n;
-	n = n - (double)left;
+	left = (long unsigned)n;
+	n = n - (long double)left;
 	str = ft_itoa(left);
 	if (len)
 	{
